+++ ExampleMod/Tiles/ExampleTorch.cs
@@ -67,7 +67,7 @@ namespace ExampleMod.Tiles
 		}
 
 		public override void PostDraw(int i, int j, SpriteBatch spriteBatch) {
-			ulong randSeed = Main.TileFrameSeed ^ (ulong)((long)j << 32 | (long)(ulong)i);
+			ulong randSeed = Main.TileFrameSeed ^ (ulong)((long)j << 32 | (long)(uint)i);
 			Color color = new Color(100, 100, 100, 0);
 			int frameX = Main.tile[i, j].frameX;
 			int frameY = Main.tile[i, j].frameY;
